Modified makefile & reorganized project structure.
Add the target "clean" and "upload-to-pic" to the makefile.
upload-to-pic uploads it to the raspberry and after that uses pickle to
program the pic. Consider making a "upload.sh" so that the makefile is
more ~general~.

Moved the driver folder inside src 'cause i mean... those are sources
too, right?