Test basic functionality for LCD.
The library was not working, for a series of problems.
The bggest one, was that i was assigning the value of the registers bits
to uint8_t type variables, because the me that was on the train after an
exam thought that was the right thing, and I didn't spot the problem
until now.
Another problem was with the registers that i was using. In particular
RA has some special functionality that I was not able to disable, but i
didn't look too mutch into it (also I'm not sure this is a real
problem).

I changed the wiring to use RC instead, and now it works (or at least it
prints text)
I also disabled 8bit mode for ease of testing because it was too mutch
hassle (considering that the only example I'm using that works on a PIC
is using 4bit and that I had to swap the wiring SO MUTCH TIMES)

I also created a file to define the pins used from the LCD, but i really
don't like this solution.