Small adjustments and testing.
Now the driver uses time.h, so it gets and receive seconds since epoch,
and all the conversions are handled internally.

Tested some functions, but I can't really test the transactions because
I don't have an oscillator for the DS1302.
If I try to send something, I get back what seems to be a random value,
but I don't know what is the behavior of the DS1302 w/o an oscillator.