Add the Main Program (tm) -- needs testing.
I'm tired.

Basically the interrupt increments an index.
If this index is less than 0, main() prints current time; if it is -ge
than 0, prints a description and an elapsed time from a date.

If the description is greater than LCD_WIDTH, at every iteration the
text is shifted to the left.
This is not implemented through scrollDisplay() because 1. i only want
to scroll one line; 2. I want the text  to appear as a circle. (It makes
sense in my head.)

I should implement the last function as a driver-side thing.

This is the second amend to add an if statement and the first one is
the same thing as the original commit i hate myself.