Unique buffer for lcd & removed atomic access to variable.
I'm using an unique buffer to print on the LCD. It wasn't strictly
necessary but it's useful to not use too much memory.

Also, I removed the functions to access the variables after disabling
the interrupts, because this would only be useful if the entire
comparisions were made in an atomic way.