Tested with oscillator -- okay to merge
Everything is working fine. The main problem in the code was with the
command byte, which i didn't understend how to format (i swear it's
not my fault, the datasheet is sheet.)

Also the oscillator i got wasn't working (and that's because i don't
know how to solder small things) so i had to mutilate an old clock.