Implemented button & timer to turn on display LCD.
Everything is working fine.
Added another interrupt (on change, but checks pin state so it's
triggered only on falling edge) to turn on the display backlight and
start a timer that turns it off.

The timer is implemented with a prescaler of 1:256, and i use a counter
to let it overflow MAX_COUNTER times. MAX_COUNTER is a macro that
expands to a formula calculating the numbers of overflows necessary to
let the backlight stay on for LCD_LED_TMR (another macro lol) seconds.

Added also a delay of DEBOUNCE_TIME before checking the pins state.