Fix bug introduced in 76b86b8c.
After an overflowing string ends it's first iteration, the LCD backlight
was forcefully turned off at every main loop. This would make it
impossible to turn it on for any significant amount of time by pressing
the interrupt button.
This is not the case anymore, because now we check if the interrupthas
been reenabled by turnOnLCD() before turing the LCD off.